+++ pipelines/folioRancher/folioDevTools/uiManagement/deployUI/Jenkinsfile
@@ -10,7 +10,7 @@ import org.folio.rest.GitHubUtility
 import org.folio.models.TenantUi
 import org.jenkinsci.plugins.workflow.libs.Library
 
-@Library('pipelines-shared-library@RANCHER-835') _
+@Library('pipelines-shared-library@RANCHER-835-v2') _
 
 properties([
     buildDiscarder(logRotator(numToKeepStr: '30')),
