+++ vars/folioStringScripts.groovy
@@ -1,3 +1,4 @@
+import groovy.json.JsonSlurperClassic
 import org.folio.Constants
 
 static String getNamespaces() {
@@ -36,4 +37,14 @@ if (installJson.getResponseCode().equals(200)) {
     }
 }
 """
+}
+
+
+static String getModuleId(String moduleName) {
+  URLConnection registry = new URL("http://folio-registry.aws.indexdata.com/_/proxy/modules?filter=${moduleName}&preRelease=only&latest=1").openConnection()
+  if (registry.getResponseCode().equals(200)) {
+    return new JsonSlurperClassic().parseText(registry.getInputStream().getText())*.id.first()
+  } else {
+    throw new RuntimeException("Unable to get ${moduleName} version. Url: ${registry.getURL()}. Status code: ${registry.getResponseCode()}.")
+  }
 }
